Rename examples `main` classes and provide and document an easy way to run them via `Gradle`
