+++ jpro-file/example/build.gradle
@@ -17,8 +17,12 @@ javafx {
     modules = ['javafx.controls', 'javafx.swing', 'javafx.media']
 }
 
-//mainClassName = 'one.jpro.platform.file.example.picker.FilePickerSample'
-mainClassName = 'one.jpro.platform.file.example.dropper.FileDropperSample'
+def examples = [
+        'text-editor' : 'one.jpro.platform.file.example.editor.TextEditorSample',
+        'file-uploader' : 'one.jpro.platform.file.example.upload.FileUploaderSample'
+]
+
+mainClassName = project.hasProperty("sample") ? examples[project.getProperties().get("sample")] : examples["text-editor"]
 
 application {
     mainClass = "$mainClassName"
