+++ jpro-file/example/src/main/java/one/jpro/platform/file/example/upload/FileSizeTableCell.java
@@ -1,4 +1,4 @@
-package one.jpro.platform.file.example.picker;
+package one.jpro.platform.file.example.upload;
 
 import javafx.geometry.Pos;
 import javafx.scene.control.ContentDisplay;
